+++ CHANGES.rst
@@ -11,6 +11,7 @@ Changelog
 - Add agency priority ordering to recent listings and listing search.
 - Agent contact form field 'phone' is now required.
 - Agent contact form field 'zipcode' is now not required anymore.
+- Use formencode's international phone validator for agent contact phone field.
 
 
 1.6.1 (2016-02-18)